REF: Move cli functions from utils.misc to cli.common - Move cli functions from utils.misc to cli.common
