+++ tiatoolbox/cli/common.py
@@ -19,7 +19,12 @@
 # ***** END GPL LICENSE BLOCK *****
 
 """Defines common code required for cli."""
+import os
+import pathlib
+
 import click
+
+from tiatoolbox.utils.misc import grab_files_from_dir, string_to_tuple
 
 cli_img_input = click.option(
     "--img-input", help="input path to WSI file or directory path"
@@ -45,3 +50,99 @@ cli_verbose = click.option(
     default=True,
     help="Print output, default=True",
 )
+
+
+def prepare_file_dir_cli(img_input, output_path, file_types, mode, sub_dirname):
+    """Prepares CLI for running code on multiple files or a directory.
+
+    Checks for existing directories to run tests.
+    Converts file path to list of file paths or
+    creates list of file paths if input is a directory.
+
+    Args:
+        img_input (str or pathlib.Path): file path to images.
+        output_path (str or pathlib.Path): output directory path.
+        file_types (str): file types to process using cli.
+        mode (str): wsi or tile mode.
+        sub_dirname (str): name of subdirectory to save output.
+
+    Returns:
+        files_all (list): list of file paths to process.
+        output_path (pathlib.Path): updated output path.
+
+    """
+    file_types = string_to_tuple(in_str=file_types)
+
+    if isinstance(output_path, str):
+        output_path = pathlib.Path(output_path)
+
+    if not os.path.exists(img_input):
+        raise FileNotFoundError
+
+    files_all = [
+        img_input,
+    ]
+
+    if os.path.isdir(img_input):
+        files_all = grab_files_from_dir(input_path=img_input, file_types=file_types)
+
+    if output_path is None and mode == "save":
+        input_dir = pathlib.Path(img_input).parent
+        output_path = input_dir / sub_dirname
+
+    if mode == "save":
+        output_path.mkdir(parents=True, exist_ok=True)
+
+    return files_all, output_path
+
+
+def prepare_model_cli(img_input, output_path, masks, file_types, mode):
+    """Prepares cli for running models.
+
+    Checks for existing directories to run tests.
+    Converts file path to list of file paths or
+    creates list of file paths if input is a directory.
+
+    Args:
+        img_input (str or pathlib.Path): file path to images.
+        output_path (str or pathlib.Path): output directory path.
+        masks (str or pathlib.Path): file path to masks.
+        file_types (str): file types to process using cli.
+        mode (str): wsi or tile mode.
+
+    Returns:
+        files_all (list): list of file paths to process.
+        masks_all (list): list of masks corresponding to input files.
+        output_path (pathlib.Path): output path
+
+    """
+    output_path = pathlib.Path(output_path)
+    file_types = string_to_tuple(in_str=file_types)
+
+    if output_path.exists():
+        raise FileExistsError("Path already exists.")
+
+    if not os.path.exists(img_input):
+        raise FileNotFoundError
+
+    if mode not in ["wsi", "tile"]:
+        raise ValueError('`mode` must be in ("wsi", "tile").')
+
+    files_all = [
+        img_input,
+    ]
+
+    if masks is None:
+        masks_all = None
+    else:
+        masks_all = [
+            masks,
+        ]
+
+    if os.path.isdir(img_input):
+        files_all = grab_files_from_dir(input_path=img_input, file_types=file_types)
+
+    if os.path.isdir(str(masks)):
+        masks_all = grab_files_from_dir(input_path=masks, file_types=("*.jpg", "*.png"))
+
+    return files_all, masks_all, output_path
